+++ app/models/appeal/excluded_appeal_contact.rb
@@ -1,7 +1,15 @@
 class Appeal::ExcludedAppealContact < ApplicationRecord
   belongs_to :appeal
   belongs_to :contact
+  validates :appeal, :contact, presence: true
+  validates :contact_id, uniqueness: { scope: :appeal_id }
+  validate :associations_have_same_account_list
 
-  validates :appeal, presence: true
-  validates :contact, presence: true
+  protected
+
+  def associations_have_same_account_list
+    return unless appeal && contact
+    return if contact.account_list_id == appeal.account_list_id
+    errors[:contact] << 'does not have the same account list as appeal'
+  end
 end