+++ app/policies/appeal/excluded_appeal_contact_policy.rb
@@ -1,0 +1,12 @@
+class Appeal::ExcludedAppealContactPolicy < ApplicationPolicy
+  def initialize(context, resource)
+    @user = context.user
+    @resource = resource
+  end
+
+  private
+
+  def resource_owner?
+    user.account_lists.exists?(id: resource.appeal.account_list_id)
+  end
+end