+++ app/serializers/appeal_contact_serializer.rb
@@ -1,0 +1,4 @@
+class AppealContactSerializer < ApplicationSerializer
+  belongs_to :appeal
+  belongs_to :contact
+end