+++ db/migrate/20170911035021_add_timestamps_to_excluded_appeal_contacts.rb
@@ -1,0 +1,7 @@
+class AddTimestampsToExcludedAppealContacts < ActiveRecord::Migration
+  def change
+    add_timestamps :appeal_excluded_appeal_contacts, default: DateTime.now
+    change_column_default :appeal_excluded_appeal_contacts, :created_at, nil
+    change_column_default :appeal_excluded_appeal_contacts, :updated_at, nil
+  end
+end